requirements.txt
cycler==0.10.0
kiwisolver==1.2.0
matplotlib==3.2.1
numpy==1.18.2
pandas==1.0.3
pymdptoolbox==4.0b3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
